requirements.txt
pillow
pillow-heif
tqdm
imageio-ffmpeg
pyinstaller
